requirements.txt
lxml==6.0.2
pillow==12.0.0
python-docx==1.2.0
python-pptx==1.0.2
typing_extensions==4.15.0
xlsxwriter==3.2.9

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# Future dependencies for assessment automation
weasyprint>=60.0  # For PDF generation from HTML
pyyaml>=6.0       # For configuration files
